Heat source, which I don't even remember what it was.
But, porting (barely) to Qt6 and CMake.

There's also making the rplidar_sdk a submodule. We need to use 1.12.0
instead of the current (2.0.0).

Cleaned up to compile with Qt6 and cmake.

But, this crashes on startup. I don't have the lidar connected. But, anyway, I think it's "ported", and now it's code and bugs. Maybe. the heat source was never completed, so that might be the issue too.

The GUI comes up. I guess the rest can be bugs.